Move the tests for the testtools modules to test_testtools
