+++ flocker/node/test/test_deploy.py
@@ -4,25 +4,12 @@
 Tests for ``flocker.node._deploy``.
 """
 
-from twisted.internet.defer import succeed
-
 from zope.interface.verify import verifyObject
 
-from ..testtools import (
-    ControllableAction, ControllableDeployer, ideployer_tests_factory,
-    EMPTY_NODE_STATE,
-)
+from ..testtools import EMPTY_NODE_STATE
 from ...testtools import TestCase
-from ...control import (
-    NodeState,
-)
 
-from .. import in_parallel
-
-from .._deploy import (
-    ILocalState, NodeLocalState,
-)
-from .istatechange import make_istatechange_tests
+from .._deploy import ILocalState, NodeLocalState
 
 
 class NodeLocalStateTests(TestCase):
@@ -48,29 +35,3 @@ class NodeLocalStateTests(TestCase):
         object_under_test = NodeLocalState(node_state=node_state)
         self.assertEqual((node_state,),
                          object_under_test.shared_state_changes())
-
-
-class ControllableActionIStateChangeTests(
-        make_istatechange_tests(
-            ControllableAction,
-            kwargs1=dict(result=1),
-            kwargs2=dict(result=2),
-        )
-):
-    """
-    Tests for ``ControllableAction``.
-    """
-
-
-class ControllableDeployerInterfaceTests(
-        ideployer_tests_factory(
-            lambda test: ControllableDeployer(
-                hostname=u"192.0.2.123",
-                local_states=[succeed(NodeState(hostname=u'192.0.2.123'))],
-                calculated_actions=[in_parallel(changes=[])],
-            )
-        )
-):
-    """
-    ``IDeployer`` tests for ``ControllableDeployer``.
-    """
